feat(user): refactor user service and handler
The user service and handler have been refactored to simplify the code.
The entity package has been removed and its functionality has been
integrated directly into the service and handler. The service now
directly uses the request objects from the v1 package, and the handler
directly uses the response objects from the v1 package. This change
reduces the amount of code and makes the service and handler easier to
understand and maintain.